requirements.txt
flask>=3.0.2
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.18.0
openai>=1.10.0
langchain>=0.1.0
langchain-openai>=0.0.2
flask-caching>=2.1.0
gunicorn>=21.2.0
geopy>=2.4.0
kaleido>=0.2.1 
folium>=0.14      